requirements.txt
PyQt5==5.15.11
PyYAML==6.0.2
jinja2==3.1.6
pyinstaller==6.15.0

# Development tools (optional)
black==23.12.1
flake8==7.0.0
isort==5.13.2
pytest==7.4.4